Moved validator, genesis and test_state from pre2p types to shared types
